+++ README.md
@@ -1,2 +1,11 @@
 # AsyncAPI Parser
 It parses AsyncAPI documents.
+
+In order to test the parser you can run:
+
+```go
+go run main.go -file <PATH_TO_DEFINITION_FILE>.yaml
+```
+
+This will output the parsed definition in json format or error in case of invalid definition file.
+
