+++ pkg/hlsp/beautifier_test.go
@@ -6,7 +6,7 @@ import (
 	"os"
 	"testing"
 
-	"github.com/asyncapi/parser/models"
+	"github.com/asyncapi/parser/pkg/models"
 	"gotest.tools/assert"
 )
