+++ pkg/parser.go
@@ -6,10 +6,6 @@ import (
 	"github.com/asyncapi/parser/pkg/hlsp"
 )
 
-var (
-	DefaultDefinitionFile = "asyncapi/2.0.0/example.yaml"
-)
-
 // Parse receives either a YAML or JSON AsyncAPI document, and tries to parse it.
 func Parse(yamlOrJSONDocument []byte) (json.RawMessage, *hlsp.ParserError) {
 	doc, err := hlsp.Parse(yamlOrJSONDocument)
@@ -17,6 +13,5 @@ func Parse(yamlOrJSONDocument []byte) (json.RawMessage, *hlsp.ParserError) {
 	if err != nil {
 		return nil, err
 	}
-
 	return doc, nil
 }
